Implement custom user checker and enhance security checks for user authentication and management

+++ src/EventListener/AuthenticationSuccessListener.php
@@ -6,16 +6,19 @@ use App\Entity\User;
 use Doctrine\ORM\EntityManagerInterface;
 use Gesdinet\JWTRefreshTokenBundle\Entity\RefreshToken;
 use Symfony\Component\Security\Core\Event\AuthenticationEvent;
+use Symfony\Component\HttpFoundation\RequestStack;
 
 class AuthenticationSuccessListener
 {
     private $entityManager;
     private $refreshTokenRepository;
+    private $requestStack;
 
-    public function __construct(EntityManagerInterface $entityManager)
+    public function __construct(EntityManagerInterface $entityManager, RequestStack $requestStack)
     {
         $this->entityManager = $entityManager;
         $this->refreshTokenRepository = $entityManager->getRepository(RefreshToken::class);
+        $this->requestStack = $requestStack;
     }
 
     /**
@@ -28,13 +31,18 @@ class AuthenticationSuccessListener
 
         // Vérifier si l'utilisateur existe et est valide
         if ($user instanceof User) {
-            // Chercher et supprimer l'ancien refresh token de cet utilisateur
-            $oldToken = $this->refreshTokenRepository->findOneBy(['username' => $user->getEmail()]);
 
-            if ($oldToken) {
-                // Supprimer l'ancien refresh token
-                $this->entityManager->remove($oldToken);
-                $this->entityManager->flush();
+            // Vérifier si la requête provient de /api/login
+            $request = $this->requestStack->getCurrentRequest();
+            if ($request && $request->getPathInfo() === '/api/login') {
+                // Chercher et supprimer l'ancien refresh token de cet utilisateur
+                $oldToken = $this->refreshTokenRepository->findOneBy(['username' => $user->getEmail()]);
+
+                if ($oldToken) {
+                    // Supprimer l'ancien refresh token
+                    $this->entityManager->remove($oldToken);
+                    $this->entityManager->flush();
+                }
             }
         }
     }
